refactor(video): split VideoPreview & VideoPlayerSection under 150 lines
VideoPreview:
- Extract playback state + handlers into useVideoPreviewPlayback hook.
- Move the staggered duration-detection effect into useRecordedVideoDuration
  hook, with its detection strategies pulled to module-scope helpers; the
  exact setTimeout schedule (and same-delay ordering) is preserved.
- Extract presentational sub-components (VideoPreviewTopBar, PlayOverlayButton,
  VideoPreviewControls, ScreenStreamVideo, PreviewReactPlayer).

VideoPlayerSection:
- Extract the player box into RecorderVideoPlayer (with PlayerTopBar and a
  module-scope duration-read helper); volume/fullscreen state stay in the
  parent and are passed down, preserving cross-source state.

Stable refs/setters added to the new hooks' dep arrays to avoid introducing
exhaustive-deps warnings (identity-stable, so no timing change).
